# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto Trading Bot API for Bybit
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategy-presets:
    get:
      operationId: listStrategyPresets
      tags: [bot]
      summary: List all strategy presets
      responses:
        "200":
          description: List of strategy presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyPreset"
    post:
      operationId: createStrategyPreset
      tags: [bot]
      summary: Create a new strategy preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyPresetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyPreset"
  /strategy-presets/{id}:
    put:
      operationId: updateStrategyPreset
      tags: [bot]
      summary: Update a strategy preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyPresetInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyPreset"
    delete:
      operationId: deleteStrategyPreset
      tags: [bot]
      summary: Delete a strategy preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
components:
  schemas:
...
    StrategyPreset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        enabled:
          type: boolean
        symbols:
          type: array
          items:
            type: string
        strategies:
          type: array
          items:
            type: string
        strategyMode:
          type: string
          enum: [AND, OR]
        positionSizeUsdt:
          type: number
        leverage:
          type: integer
        maxPositions:
          type: integer
        stopLossUsdt:
          type: number
        takeProfitUsdt:
          type: number
        averagingEnabled:
          type: boolean
        averagingThresholdPercent:
          type: number
        maxAveragingCount:
          type: integer
        averagingAmountUsdt:
          type: number
        timeframe:
          type: string
        volumeProfileParams:
          type: object
        fibonacciParams:
          type: object
        orderBlockParams:
          type: object
        rsiParams:
          type: object
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - enabled
        - symbols
        - strategies
        - strategyMode
        - positionSizeUsdt
        - leverage
        - maxPositions
        - stopLossUsdt
        - takeProfitUsdt
        - averagingEnabled
        - averagingThresholdPercent
        - maxAveragingCount
        - averagingAmountUsdt
        - timeframe
        - createdAt
        - updatedAt
    StrategyPresetInput:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        symbols:
          type: array
          items:
            type: string
        strategies:
          type: array
          items:
            type: string
        strategyMode:
          type: string
          enum: [AND, OR]
        positionSizeUsdt:
          type: number
        leverage:
          type: integer
        maxPositions:
          type: integer
        stopLossUsdt:
          type: number
        takeProfitUsdt:
          type: number
        averagingEnabled:
          type: boolean
        averagingThresholdPercent:
          type: number
        maxAveragingCount:
          type: integer
        averagingAmountUsdt:
          type: number
        timeframe:
          type: string
        volumeProfileParams:
          type: object
        fibonacciParams:
          type: object
        orderBlockParams:
          type: object
        rsiParams:
          type: object
      required:
        - name